Added option to disable Capirca ACL generation completely.
Capirca filters are generated and added to device configurations
outside the scope of the templates defined for device roles.

For some lab testing I was doing I wanted to use the Netbox plugin,
and pull data representing actual WMF network topology.  However,
as I was using crpd-based containers for simulation, I used
different top-level configuration tempaltes to generate the "device"
configs·

An issue I hit was that with the real WMF data the Capirca filters
were generated every time, and added to the resulting JunOS device
config.

This was problematic with crpd as it uses a Linux data-plane and does
not support JunOS style filters, i.e. crpd has no "firewall"
configuration section.

Adding this optional attribute to config.yaml, to disable Capirca filter
generation even if device config data has them.

I'm also proposing another slight change, to the default for the
capirca 'netbox_definitions' option.  This was defaulting to false,
which meant that unless the 'capirca' section was defined, and
it was set to 'true', config generation would fail in WMF.

The change means that if Netbox is being used it's assumed to also
be the source for definitions.  If Netbox is not being used it will
default to getting the data from files.  If Netbox is being used,
but the user wants definitions to come from files, then the
configuration option needs to be set to 'false'.

Change-Id: I5e058e45908f5f5b193c1b3e5e9a7aa93f9a8f86