NetboxInventory: use GraphQL and save ~30s at each run
From ~38s to ~3s.

Reused John's code from sync-netbox-hiera.py <3

Remove VC specific code as the VC master device is the only one to have
a FQDN, so it just works.
Initialise a Netbox (pynetbox) object only when we need it.
Not sure if there is a better location for _gql_execute(). Maybe not
make it private so it can be re-used by the wmf-netbox plugin?

Tests from Ifb0540cd0f94caf5ebd25c64b59cf73dd31d5bd6

Bug: T310577
Change-Id: Id88ed0c46cac9e01e63053392e4e1fd46e372807