Capirca: deal with scenario when netbox script has never run
If the capirca netbox script has never been run (as opposed to run
but not yet completed, or run but has an error), homer currenlty
exits with a generic 'AttributeError: object has no attribute
"status"'.

This is quite common when running homer during testing against
netbox-next.  Which I do just infrequently enough to forget about
this issue.

This patch will catch the problem by checking for the 'completed'
key in the results, rather than 'status' key, which is not
present if it has never run.  I checked and if the script is
currently running the 'completed' key is also null (even if it
has sucessfully run before), so we can use the single check.

Change-Id: I2ae7c63dcd44dd8bca9e3507741a171ac5fc613d