JunOS: pass ignore_warnings list to diff() and rollback() functions
As described on task adjusting the chassis port-block stanza causes a
warning when we trigger the PyEz device.Config.load/diff/rollback()
functions.

We have an appropriate sub-string in our default config.yaml for
Homer, which mainifests as ignore_warning, however we were not
passing it to the diff or rollback functions, which was triggering
an RpcError from PyEz when the block config changed.

Additionally I changed what we pass to load(), passing our defined
list of what to ignore, previously we were just passing True which
would accept all errors.

Works as expected:

https://phabricator.wikimedia.org/P79742

Bug: T400261
Change-Id: Ia88e5b927592fd654587543645d201951dc87931